requirements.txt
selenium==4.23.1
pytest==8.3.2
allure-pytest==2.13.5